Use a single TCP server port in runner
The shim is now passed a shim ID which allows us to dispatch.
Previously, we dispatched with a distinct TCP server port per test, but
this seems to be causing port exhaustion.

As with https://boringssl-review.googlesource.com/c/boringssl/+/60885,
this changes the runner/shim protocol, but not the shim/handshaker
protocol. See that CL for discussion.

Change-Id: Ia22d1bc0bcda1b1869bce4630138538c30f02668