Remove OPENSSL_TRUSTY ifdef in crypto/bio/file.c
Trusty has FILE and BIO, so it's weird that the combination is ifdef'd
out. It's further weird that we still have functions that call into the
excluded functions.

Just build the file on Trusty. It seems to work just fine. Trusty uses
musl these days (b/110161494).

Update-Note: If the build breaks on Trusty in this file, it's probably
this CL's fault. But I tested it and it seems to be fine.

Change-Id: Ic85402a04739f891045451f3adc8204698522924