Use sources.cmake for the test_support library
As part of this, align the generated and standalone builds in how
crypto/test/gtest_main.cc is added. Since not all test targets
(urandom_test) include gtest_main.cc, and the generated one just
manually adds it to the file lists, just put it into the file lists
ahead of time. That way we don't need to synchronize the dependency
information with the generated build.

Bug: 542
Change-Id: I235e0ccd0432f4b380a92b265ede35eb8a6a6e36