Don't include stdalign.h in refcount.c
Instead, rely on internal.h to either include the header or polyfill it.
On Windows, we don't (yet) require C11, so we can't rely on the header
being directly includable. Though given it took a couple months to
notice this, it's clear the non-C11 path is pretty much untested, so we
need to get rid of it.

Bug: 624
Change-Id: I86a6961c93161c3adfacd374affb8bfb2be0a542