Move file_test_gtest.cc to CRYPTO_TEST_SOURCES for now
file_test_gtest.cc depends on GetTestData which, for now, only exists in
crypto_test because of how the crypto_test_data machinery works. Ideally
that would be fixed but, for now, move the file out of test_support.

Things were previously building because we were relying on the static
linker to drop the file in the other targets, before the linker would
notice the undefined symbol. I'm not sure what's different about
Chromium, but somehow Chromium is sensitive to this.

Bug: 542
Change-Id: I38c7719e7c3f27bcfeeab9693fcd43fbd0300799