Add combined source lists to the other generators
This doesn't remove the old ones yet as we'll need a corresponding
downstream change. This does all of them but Soong. I think, for Soong,
we can apply the change unilaterally in the source list, but I'm holding
off on testing that until I've gotten Android updated to a more recent
BoringSSL.

Bug: 542
Change-Id: I30921ac251f76e3ecb50df9ff757fec4031e7dbc