Unconditionally include BIO_set_fd and BIO_get_fd
These functions are just wrappers over BIO_ctrls, shared between the fd
and socket BIOs. Though we don't currently support one, it is
conceivable that there would be a platform with socket BIOs but not fd
BIOs. In that case, the BIO_get_fd function would still be useful to
implement SSL_get_rfd. (And someone could conceivably implement it in
another BIO.)

Bug: 629
Change-Id: I7ac3561f76af86af32d70b1c6265c4caeaecb129