Use a stub fopen implementation when OPENSSL_NO_FILESYSTEM is set
Detecting errors (i.e. fs-less platforms using fs-only APIs) at compile
time is generally preferable to doing so at runtime, so
https://boringssl-review.googlesource.com/c/boringssl/+/61726 opted to
remove the APIs altogether on applicable targets.

However, Trusty uses rust-openssl somewhere and rust-openssl binds a
bunch of filesystem-dependent APIs unconditionally. To keep that
working, switch to a stub fopen when OPENSSL_NO_FILESYSTEM is set. We
effectively model a platform where the filesystem "exists", but is
empty. Upstream OpenSSL similarly has OPENSSL_NO_STDIO still define the
file BIO (unlike the socket BIO, which is excluded), but in a stub form.

For now, I've kept by_dir.c and by_file.c excluded on Trusty, even
though it's not reflected in the header and they no longer need to be
excluded. It also means Trusty is actually relying on dead code
elimination removing the symbol references to the nonexistent symbols.
But in the interest of getting the current Android update through, I'm
trying to align with the old state of Trusty as much as possible. Once
the update is through, I'll switch it to something more self-consistent.

Bug: 629
Change-Id: I4f20d872a7cde863d21c78090f270b77b03545fa