Switch bssl-sys to build with bindgen API, not CLI
Bindgen API offers some knobs to fix issues that the CLI
does not. The API is the recommended way to use bindgen:
https://rust-lang.github.io/rust-bindgen/library-usage.html.

The cmake configuration for bssl-sys now relies on cargo instead
of the bindgen command to output the Rust bindings wrapper, but
it attempts to do the least work possible by calling `cargo check`
instead of `cargo build` and wrapping a `_bindgen_only` feature around
the entire `lib.rs` source file that suppresses actual compilation
if the feature is enabled.

Change-Id: I0df79f2af8903a7d2177a0d1e1a80337888a2357