Remove the remaining old-style asm lists
I think there are a couple projects remaining that still use the old
lists here, but they're in repositories we don't spend as much time in,
and it should be straightforward for them to update when they get here.

Removing these should put us in a good place to check in pre-generated
asm lists. While I'm here, fix a few typos in TODOs I previously added.

Update-Note: If you're one of those projects and have trouble switching
to the new lists, let us know.

Bug: 542
Change-Id: I57559bafc85eceacc7a237e2f29db6eaf492a8cb