Silence warn_unused_result warning on write() call
Frustratingly, simply writing the standard (void)write(...) does not
work because GCC is broken and intentionally leaves the warning enabled
there. This does not comply with the now standard semantics for
nodiscard.
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66425

Instead, what seems to work is to assign it to a variable and then
(void) the variable.

Fixed: 644
Change-Id: Ic418b4185aeae1a9ca424c45a05af063e8d50255