Explicitly mark saved registers with .cfi_restore
Although it works without these (we just refer the unwinder to the red
zone), older versions of libunwind seem to have a bug that cause it to
flakily fail to restore rbx without this. I've attempted to bisect the
problem, but the issue is very flaky and I've failed to find the culprit
four times now, so just give up and work around it. Explicit restores
match what we do in other files.

Hopefully this will clear some issues tha fiat-crypto's CI are running
into.

Change-Id: I6a19679a37cad8e93e6dee554b6a9b3b9b4bbe4a